OBD2_Live_Monitor/requirements.txt
altair==4.2.2
argparse-addons==0.12.0
attrs==22.2.0
bitstruct==8.17.0
blinker==1.5
cachetools==5.3.0
cantools==38.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
crccheck==1.3.0
decorator==5.1.1
diskcache==5.4.0
dnspython==2.3.0
entrypoints==0.4
getmac==0.9.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.1.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
msgpack==1.0.5
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.14.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
pymongo==4.3.3
Pympler==1.0.1
pynmea2==1.19.0
pyrsistent==0.19.3
pyserial==3.5
python-can==4.1.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.20.0
tenacity==8.2.2
textparser==0.24.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wrapt==1.15.0
zipp==3.15.0
